+++ .changeset/mean-taxis-tie.md
@@ -3,4 +3,5 @@
 ---
 
 - add support for custom tailwind prefix
+- add `extend` and `create` to `lib/utils.tsx`
 - minify build
