+++ apps/www/content/docs/components/changelog.mdx
@@ -10,6 +10,10 @@ Use the [CLI](https://platejs.org/docs/components/cli) to install the latest ver
 
 ## December 2023 #6
 
+### December 23 #6.2
+
+- add `extend` and `create` to `lib/utils.tsx`
+
 ### December 10 #6.1
 
 - `image-element`: wrap the component with `withHOC(ResizableProvide, ...)`
