+++ packages/cli/src/utils/templates.ts
@@ -1,8 +1,40 @@
-export const UTILS = `import { type ClassValue, clsx } from 'clsx'
+export const UTILS = `import {
+  createElement,
+  ElementRef,
+  forwardRef,
+  ForwardRefExoticComponent,
+} from 'react';
+import { type ClassValue, clsx } from 'clsx'
 import { twMerge } from 'tailwind-merge'
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
+}
+
+export function extend<T extends { className?: string }>(
+  Component: ForwardRefExoticComponent<T>,
+  defaultProps: T
+) {
+  return forwardRef<ElementRef<typeof Component>, T>(
+    function ExtendComponent(props, ref) {
+      return (
+        <Component
+          ref={ref}
+          {...defaultProps}
+          {...props}
+          className={cn(defaultProps.className, props.className)}
+        />
+      );
+    }
+  );
+}
+
+export function create<T extends keyof HTMLElementTagNameMap>(tag: T) {
+  return forwardRef<HTMLElementTagNameMap[T], JSX.IntrinsicElements[T]>(
+    function CreateComponent(props, ref) {
+      return createElement(tag, { ...props, ref });
+    }
+  );
 }
 `;
